refactor: [Person] SSR 전환 및 hook 패턴 적용 - 비지니스 로직, ui 로직 분리
